+++ packages/core/src/decoder/index.js
@@ -1,2 +1,3 @@
 export { decodeAction } from './actionDecoder.js';
 export { simulateAction } from './txSimulator.js';
+export { balancesFromSdk } from './balanceAdapter.js';